Split out image tasks
