+++ concordia/admin/views.py
@@ -30,10 +30,8 @@ from concordia.models import (
 from exporter.tabular_export.core import export_to_csv_response, flatten_queryset
 from exporter.views import do_bagit_export
 from importer.models import ImportItem, ImportItemAsset, ImportJob
-from importer.tasks import (
-    fetch_all_urls,
-    redownload_image_task,
-)
+from importer.tasks import fetch_all_urls
+from importer.tasks.images import redownload_image_task
 from importer.tasks.items import import_items_into_project_from_url
 from importer.utils import slurp_excel
